Refactored predict-api as explainability addon + accept batch requests
