+++ addons/explainability/pom.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>addons</artifactId>
+        <groupId>org.kie.kogito</groupId>
+        <version>1.0.0-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>explainability</artifactId>
+    <name>Kogito :: Add-Ons :: Explainability</name>
+
+    <packaging>pom</packaging>
+
+
+    <modules>
+        <module>explainability-common</module>
+        <module>explainability-springboot-addon</module>
+        <module>explainability-quarkus-addon</module>
+    </modules>
+
+
+</project>
