--- addons/tracing/tracing-decision-api/src/main/java/org/kie/kogito/tracing/decision/event/explainability/ModelIdentifier.java
@@ -1,55 +1,0 @@
-/*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.kogito.tracing.decision.event.explainability;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class ModelIdentifier {
-
-    public static final String RESOURCE_ID_SEPARATOR = ":";
-
-    @JsonProperty("resourceType")
-    private String resourceType;
-
-    @JsonProperty("resourceId")
-    private String resourceId;
-
-    public ModelIdentifier(){}
-
-    public ModelIdentifier(String resourceType, String resourceId){
-        this.resourceType = resourceType;
-        this.resourceId = resourceId;
-    }
-
-    public String getResourceType() {
-        return resourceType;
-    }
-
-    public void setResourceType(String resourceType) {
-        this.resourceType = resourceType;
-    }
-
-    public String getResourceId() {
-        return resourceId;
-    }
-
-    public void setResourceId(String resourceId) {
-        this.resourceId = resourceId;
-    }
-}
