--- addons/tracing/tracing-decision-api/src/main/java/org/kie/kogito/tracing/decision/event/explainability/PredictInput.java
@@ -1,55 +1,0 @@
-/*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.kogito.tracing.decision.event.explainability;
-
-import java.util.Map;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class PredictInput {
-
-    @JsonProperty("modelIdentifier")
-    private ModelIdentifier modelIdentifier;
-
-    @JsonProperty("request")
-    private Map<String, Object> request;
-
-    public PredictInput(){}
-
-    public PredictInput(ModelIdentifier modelIdentifier, Map<String, Object> request) {
-        this.modelIdentifier = modelIdentifier;
-        this.request = request;
-    }
-
-    public ModelIdentifier getModelIdentifier() {
-        return modelIdentifier;
-    }
-
-    public void setModelIdentifier(ModelIdentifier modelIdentifier) {
-        this.modelIdentifier = modelIdentifier;
-    }
-
-    public Map<String, Object> getRequest() {
-        return request;
-    }
-
-    public void setRequest(Map<String, Object> request) {
-        this.request = request;
-    }
-}
