+++ addons/tracing/tracing-decision-common/pom.xml
@@ -24,6 +24,11 @@
     </dependency>
 
     <dependency>
+      <groupId>org.kie.kogito</groupId>
+      <artifactId>explainability-common</artifactId>
+    </dependency>
+
+    <dependency>
       <groupId>org.junit.jupiter</groupId>
       <artifactId>junit-jupiter-engine</artifactId>
       <scope>test</scope>
