+++ addons/tracing/tracing-decision-common/src/main/java/org/kie/kogito/tracing/decision/DecisionTracingListener.java
@@ -27,13 +27,14 @@ import org.kie.dmn.api.core.event.DMNEvent;
 import org.kie.dmn.api.core.event.DMNRuntimeEventListener;
 import org.kie.kogito.tracing.decision.event.evaluate.EvaluateEvent;
 
+import static org.kie.kogito.explainability.ExplainabilityResourceExecutor.SKIP_TRACING;
+
 /**
  * This class must always have exact FQCN as <code>org.kie.kogito.tracing.decision.DecisionTracingListener</code>
  * for code generation plugins to correctly detect if this addon is enabled.
  */
 public class DecisionTracingListener implements DMNRuntimeEventListener {
 
-    public final static String SKIP_TRACING = "skipTracing";
     private Consumer<EvaluateEvent> eventConsumer;
 
     public DecisionTracingListener(Consumer<EvaluateEvent> eventConsumer) {
