+++ addons/tracing/tracing-decision-quarkus-addon/pom.xml
@@ -27,6 +27,11 @@
     </dependency>
 
     <dependency>
+      <groupId>org.kie.kogito</groupId>
+      <artifactId>explainability-quarkus-addon</artifactId>
+    </dependency>
+
+    <dependency>
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-smallrye-reactive-messaging-kafka</artifactId>
     </dependency>
