--- addons/tracing/tracing-decision-quarkus-addon/src/test/java/org/kie/kogito/tracing/decision/Constants.java
@@ -1,24 +1,0 @@
-/*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.kogito.tracing.decision;
-
-public class Constants {
-    public final static String MODEL_RESOURCE = "/Traffic Violation.dmn";
-    public final static String MODEL_NAMESPACE = "https://github.com/kiegroup/drools/kie-dmn/_A4BCA8B8-CF08-433F-93B2-A2598F19ECFF";
-    public final static String MODEL_NAME = "Traffic Violation";
-    public final static String TEST_EXECUTION_ID = "testExecutionId";
-}
