+++ addons/tracing/tracing-decision-quarkus-addon/src/test/java/org/kie/kogito/tracing/decision/DecisionModelResourcesProviderMock.java
@@ -26,9 +26,7 @@ import org.kie.internal.decision.DecisionModelResourcesProvider;
 import org.kie.kogito.decision.DecisionModelType;
 import org.kie.kogito.dmn.DecisionModelRelativeResource;
 
-import static org.kie.kogito.tracing.decision.Constants.MODEL_NAME;
-import static org.kie.kogito.tracing.decision.Constants.MODEL_NAMESPACE;
-import static org.kie.kogito.tracing.decision.Constants.MODEL_RESOURCE;
+import static org.kie.kogito.tracing.decision.QuarkusDecisionTracingTest.*;
 
 @Mock
 public class DecisionModelResourcesProviderMock implements DecisionModelResourcesProvider {
