+++ addons/tracing/tracing-decision-springboot-addon/pom.xml
@@ -31,6 +31,11 @@
     </dependency>
 
     <dependency>
+      <groupId>org.kie.kogito</groupId>
+      <artifactId>explainability-springboot-addon</artifactId>
+    </dependency>
+
+    <dependency>
       <groupId>org.springframework.kafka</groupId>
       <artifactId>spring-kafka</artifactId>
     </dependency>
